+++ src/imaging/FWCameraUtils.h
@@ -36,7 +36,7 @@
 namespace avg {
 
 dc1394video_mode_t getCamMode(IntPoint size, PixelFormat pf);
-dc1394framerate_t getFrameRateConst(double frameRate);
+dc1394framerate_t getFrameRateConst(float frameRate);
 CameraFeature featureIDToEnum(dc1394feature_t feature);
 dc1394feature_t getFeatureID(CameraFeature feature);
 IntPoint getFrameSizeFromVideoMode(dc1394video_mode_t mode);
